support sorting table rows on frontend
