+++ coveralls.json
@@ -1,6 +1,6 @@
 {
   "coverage_options": {
-    "minimum_coverage": 79.7
+    "minimum_coverage": 89.7
   },
   "skip_files": []
 }
