+++ lib/nfl_rushing_web/live/rushing_statistics_live.html.leex
@@ -5,7 +5,16 @@
       <button type="submit" phx-disable-with="Searching...">Search</button>
     </form>
   </div>
+  <div class="column">
+    <% url = "/csv?search=#{@search}&sort=#{@sort}&order=#{@order}" %>
+    <a download href="<%= url %>">Export CSV</a>
+  </div>
 </section>
 <section name="statistics" class="row">
-  <%= live_component NflRushingWeb.StatisticsComponent, rushing_statistics: @rushing_statistics %>
+  <table style="width:100%">
+    <%= live_component NflRushingWeb.StatisticsHeaderComponent, search: @search, sort: @sort, order: @order %>
+    <%= for stat <- @rushing_statistics do %>
+      <%= live_component NflRushingWeb.StatisticsRowComponent, rushing_statistics: stat %>
+    <% end %>
+  </table>
 </section>
